refactor: changing tooglethemeprovider name to avoid breaking the pipeline
